use a hook for getting collections
